Finding stops now works well.
