--- Map matching/computing azimuth_formula_for_QGIS.txt
@@ -1,2 +1,0 @@
-(atan((xat(-1)-xat(0))/(yat(-1)-yat(0)))) * 180/3.14159 + (180 *(((yat(-1)-yat(0)) < 0) + (((xat(-1)-xat(0)) < 0 AND (yat(-1) - yat(0)) >0)*2)))
-
