+++ map_matching/aequilibrae/__init__.py
@@ -1,0 +1,26 @@
+"""
+/***************************************************************************
+ AequilibraE - www.aequilibrae.com
+ 
+    Name:        QGIS plugin initializer
+                              -------------------
+        begin                : 2014-03-19
+        copyright            : AequilibraE developers 2014
+        Original Author: Pedro Camargo pedro@xl-optim.com
+        Contributors: 
+        Licence: See LICENSE for https://github.com/AequilibraE/AequilibraE
+ ***************************************************************************/
+
+"""
+
+# This portion of the script initializes the plugin, making it known to QGIS.
+import sys
+sys.dont_write_bytecode = True
+
+from .graph import Graph        # We import the graph
+#from .assignment import *
+
+
+# We import the algorithms
+import platform
+plat = platform.system()
