--- src/__init__.py
@@ -1,2 +1,0 @@
-from trip import Trip
-from network import Network
